This work is not complete yet, but I need food.
Always use list of modules rather than dicts.  We end up with some list
comps to make lists of the data we are interested in at the time, but
its logically simpler.

resolveVersions() now does the hard work of matching up kmods to kernels.
Use its tables in our magic.

installAll has been written.  Working on porting pinKernels next.

No this is not tested and will not work.


git-svn-id: https://linuxczar.net/svn/slack/trunk/fedorakmod@786 8bb72887-34d4-0310-a9bb-bbd9c5be255b