This contains an old fix to that throws the RPM flag to allow installs of an older kmod (rpm evr wise) than one that's currently installed.  (Say, you needed to grab an older kernel or a kmod for an older kernel becomes available.)
I have also hacked on pinKernels() a bit to cover the case where we've
asked to pin the kernel but we don't have the currently running kernel's
RPM package installed.  In that case we chose the highest EVR installed
kernel package to use as our reference for the kmods we watch for.


git-svn-id: https://linuxczar.net/svn/slack/trunk/fedorakmod@800 8bb72887-34d4-0310-a9bb-bbd9c5be255b